260430-fairy-story-agent/pyproject.toml
[project]
name = "260430-fairy-story-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.31.1",
    "langgraph>=1.1.10",
    "litellm>=1.83.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
